ai-worker/requirements.txt
redis==5.0.1
pydantic==2.5.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
